weather_api/pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = "\"Making a weather app using public api\""
authors = ["Glen llaci"]
license = "glen"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.2"
uuid = "1.30"
authlib = "^1.2.1"
ipython = "8.11.0"
django-extensions = "^3.2.1"
Markdown = "^3.5.1"
pillow = "9.3.0"
requests = "^2.31.0"
django-solo = "^2.1.0"
python-dateutil = "^2.8.2"
pytz = "^2023.3.post1"
python-dotenv = "^1.0.0"
cryptography = "^41.0.5"
ipdb = "^0.13.13"
djangorestframework = "^3.14.0"
django-stdimage = "^6.0.2"
stripe = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
